Move tests for Group::create into separate test case
